requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.9.0
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.4
faiss-cpu==1.8.0
fastapi==0.115.12
filelock==3.13.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.22.2
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
json-html==0.0.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain-community==0.0.32
langchain-core==0.1.42
langsmith==0.1.47
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
orjson==3.10.0
packaging==23.2
pandas==2.2.1
pillow==10.3.0
pluggy==1.3.0
protobuf==4.25.3
pyarrow==15.0.2
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
PyPDF2==3.0.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.46.2
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
transformers==4.39.3
txtai==7.0.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.34.2
watchdog==4.0.0
yarl==1.9.4
